requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
requests>=2.31.0
python-dotenv>=1.0.0

# Trading and financial data
ib-insync>=0.9.85
yfinance>=0.2.65

# Data processing and utilities
urllib3>=2.0.0
certifi>=2023.0.0

# Optional: For enhanced logging and monitoring
psutil>=5.9.0

# Development and testing (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0